requirements.txt
flask
flask-cors
mysql-connector-python
pandas
numpy
scikit-learn
joblib
polygon-api-client
python-dotenv
gunicorn